+++ src/hooks/use-mount.ts
@@ -1,7 +1,0 @@
-const mount = (callback: Function) => {
-  useEffect(() => {
-    callback && callback()
-  }, [])
-}
-
-export default mount